requirements.txt
torch==2.2.2
torchvision==0.17.2
scikit-learn==1.3.2
seaborn==0.13.2
matplotlib==3.7.3
tqdm==4.67.1
numpy==1.24.3
pillow==11.1.0
requests==2.32.3
